Map, Lighting, Handling and Input updates
Map:
Major changes to the map, including a new oval section with a banked corner and other updates to the scene to make it harder to cheat and, to make the map feel more open and less cramped.

Lighting:
The first version of a day night cycle has been added.

Handling:
The handling has now been changed to almost eliminate flips and the handling now feels sharper, but you can still slide if you want.

Working on adding gears.

Inputs:
The game should now support any game pad for inputs.